# add iter pool to reduce allocations