@tidwall 2023-11-18
It makes sense to add pooling for iterators.
Your code is has a (slightly) breaking change with adding the pointer. Not sure if that's a big deal in practice though.
Also, the Pools can't be copied so the Copy function may have an issue. Perhaps creating the pool as a pointer type `&Pool{}` will fix this. 
I'll need to review further, but my time is a bit limited atm, hope to check back soon. thx.